feat(crates): add tofu crate
This crate contains logic and data structures that can be used to
represent the state of a configuration. This is done through the
`TerraformStateProvider` trait. An implementation that stores state in
memory is provided.
