+++ Cargo.toml
@@ -3,17 +3,24 @@ members = ["crates/*"]
 resolver = "2"
 
 [workspace.package]
-edition = "2021"
-homepage = "https://github.com/justinrubek/thoenix"
-license = "MIT"
-repository = "https://github.com/justinrubek/thoenix"
 version = "0.2.0"
+edition = "2021"
+license = "MIT"
+homepage = "https://github.com/justinrubek/thoenix"
+repository = "https://github.com/justinrubek/thoenix"
+authors = ["Justin Rubek"]
 
 [workspace.dependencies]
 serde_json = "1"
+thiserror = "1.0.38"
+tracing = "0.1.37"
 
 [workspace.dependencies.serde]
 version = "1"
 features = ["derive"]
+
+[workspace.dependencies.tokio]
+version = "1"
+features = ["full"]
 
 [profile.release.package]
