+++ crates/http/Cargo.toml
@@ -14,6 +14,7 @@ hyper = "0.14.24"
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = "1.0.38"
+thoenix-tofu = { path = "../tofu" }
 tracing = "0.1.37"
 
 [dependencies.tokio]
