+++ crates/tofu/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "thoenix-tofu"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+authors = { workspace = true }
+
+[dependencies]
+serde = { workspace = true }
+thiserror = { workspace = true }
